Refactored the Atomix specific instance of ServerImplementationTest
